change collision detection to work with hitbox size instead of sprite tile size

+++ src/main/java/com/codecool/platformer/utils/HelpMethods.java
@@ -4,7 +4,7 @@ import com.codecool.platformer.constants.GameProperties;
 
 public class HelpMethods {
 
-    public static boolean canMoveHere(float x, float y, int width, int height, int[][] levelData) {
+    public static boolean canMoveHere(float x, float y, float width, float height, int[][] levelData) {
         if (!isSolid(x, y, levelData)) {
             if (!isSolid(x + width, y + height, levelData)) {
                 if (!isSolid(x + width, y, levelData)) {
